Renamed test files
