+++ packages/jest-expo/tests/__tests__/jest-test.js
@@ -1,11 +1,11 @@
 describe(`jest-expo/universal`, () => {
-  it(`[moduleFileExtensions] default`, () => {
-    const { OS } = require('../foo');
+  it(`resolves to default extensions`, () => {
+    const { OS } = require('../default-extension');
     const { Platform } = require('@unimodules/core');
     expect(OS).toContain(Platform.OS);
   });
-  it(`[moduleFileExtensions] fallback`, () => {
-    const { OS } = require('../bar');
+  it(`resolves to fallback extensions`, () => {
+    const { OS } = require('../fallback-extension');
     const { Platform } = require('@unimodules/core');
     expect(OS).toContain(Platform.OS);
   });
