+++ packages/jest-expo/tests/default-extension.android.js
